move database classes to separate folder move rotatask to adapters
