+++ app/src/main/java/in/antaragni/ant/adapters/RotaTask.java
@@ -1,4 +1,4 @@
-package in.antaragni.ant;
+package in.antaragni.ant.adapters;
 
 
 import android.content.Context;
