+++ app/src/main/java/in/antaragni/ant/datahandler/DatabaseOpenHelper.java
@@ -1,4 +1,4 @@
-package in.antaragni.ant;
+package in.antaragni.ant.datahandler;
 
 import android.content.Context;
