+++ app/src/main/java/in/antaragni/ant/fragments/MapFragment.java
@@ -4,7 +4,6 @@ package in.antaragni.ant.fragments;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,13 +13,12 @@ import android.widget.Button;
 import android.widget.Spinner;
 
 
-import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.SupportMapFragment;
 
-import in.antaragni.ant.DatabaseAccess;
+import in.antaragni.ant.datahandler.DatabaseAccess;
 import in.antaragni.ant.R;
-import in.antaragni.ant.RotaTask;
+import in.antaragni.ant.adapters.RotaTask;
 import in.antaragni.ant.datamodels.Venue;
